+++ packages/cloud-cognitive/src/global/js/utils/hasValidChildType.js
@@ -6,14 +6,14 @@
  */
 
 // Custom PropType validator which checks and ensures that the children of the create component are indeed all CreateStep components.
-export const isValidChildren =
-  ({ componentName, stepType }) =>
+export const hasValidChildType =
+  ({ componentName, childType }) =>
   ({ children }) => {
     children.length > 1 &&
       children.map((child) => {
-        if (child && child.props && child.props.type !== stepType) {
+        if (child && child.props && child.props.type !== childType) {
           throw new Error(
-            `Each child of ${componentName} is required to be a ${stepType}. Please remove the HTML element, or wrap it around the ${stepType} component.`
+            `Each child of ${componentName} is required to be a ${childType}. Please remove the HTML element, or wrap it around the ${childType} component.`
           );
         }
         return;