+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.70.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.70.0...@carbon/ibm-cloud-cognitive@0.70.1) (2021-08-17)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+# [0.70.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.69.7...@carbon/ibm-cloud-cognitive@0.70.0) (2021-08-16)
+
+
+### Features
+
+* **CreateSidePanel:** release create side panel component ([#1153](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1153)) ([5b39d10](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/5b39d1095f1a52d020985a9c8cf1cebe89c5071a))
+
+
+
+
+
 ## [0.69.7](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.69.6...@carbon/ibm-cloud-cognitive@0.69.7) (2021-08-13)